Complete Function Cache codewars kata
This was a pretty simple and naive implementation. A more interesting one would be one that supported any number of nested operations, and it probably wouldn't take much to expand my solution to include that. Each operation function should test if the one argument it receives is actually a number, and if it is, create an equation object with that number as the right operand (b property).