Complete Pick Peaks codewars kata
This was relatively straightforward, but was still helpful for getting me to practice simplifying the rules I use when dealing with a situation like this one. I started off with more complicated if conditions - if arr[i] > arr[i-1] && arr[i] > arr[i+1], etc. That way required less thinking, since I could specify each potential situation, but it would have led to lots more if blocks, and thus harder-to-read and -to-maintain code. Then I spent a bit more time thinking about how to solve the problem, and came up with this simpler solution. (It could probably still be simplified some.)