[WIP] Last Digit of a Huge Number (Code Wars kata) - need to preserve more sigdigs
This passes almost all the tests, and in fact once actually passed ALL the tests. But it isn't complete yet, because the tests are somewhat randomized - the time I passed, there just didn't happen to be any tests that my function would fail. I'm committing here so I'll have a record of this implementation, since I'm about to restructure several lines of code in ways that won't be quick to reverse.