Create naive approach and complete test framework
I've created the naive prisoner approach of always declaring that you're the final prisoner if you're visiting the room on the nth day, where there are n prisoners. I've also completed the test framework - it generates a bunch of random sequences of prisoners, and then tries out a prisoner strategy on a sequence, and returns an array of final results of whether that strategy worked on each of the sequences.