Create functions that can go into prisoner strategies
I'm planning to recursively create different combinations of strategies prisoners can use to determine whether to turn the light on or off, and whether to assert that they're the final prisoner. These functions will help me construct those various strategies. For example, a strategy might be:
Set the light to on if it is already on OR if I have visited an even number of times
Assert that I am the final prisoner if it is not an even day AND if the light is off AND if this is my first visit.